feat: Add new UI components and game mechanics for Homeless Hero
- Implement FightModal for combat interactions, displaying enemy and player stats.
- Create HpBar component for health representation.
- Add NpcDialog for NPC interactions and options.
- Develop ShopModal for managing player inventory and shop transactions.
- Introduce StatusModal for displaying player status and quests.
- Build WorldView to manage game state and navigation.
- Establish useRpgGame hook for game logic and state management.
- Add PatternMemory game with grid-based memory mechanics and scoring system.
- Create smallHeader for navigation and user profile display.
- Implement AchievementContext for managing player achievements and rewards.

+++ MartinyArcadeApi/Data/Models/GameSession.cs
@@ -8,7 +8,7 @@ public class GameSession
     public string GameKey { get; set; } = string.Empty;
 
     public int Score { get; set; }
-    public int XpEarned { get; set; }
+    public long XpEarned { get; set; }
 
     public DateTime PlayedAt { get; set; } = DateTime.UtcNow;
