+++ MartinyArcadeApi/Data/Models/GameStats.cs
@@ -5,7 +5,7 @@ public class UserGameStats
     public Guid UserId { get; set; }
     public string GameId { get; set; } = string.Empty;
 
-    public int TotalXP { get; set; }
+    public long TotalXP { get; set; }
     public int TotalPlays { get; set; }
     public DateTime LastPlayed { get; set; }
