+++ MartinyArcadeApi/Data/Models/UserProfile.cs
@@ -5,7 +5,7 @@ public class UserProfile
 {
     [Key]
     public Guid UserId { get; set; }
-    public int TotalXP { get; set; }
+    public long TotalXP { get; set; }
     public DateTime LastUpdated { get; set; }
 
     public User User {get; set; } = null!;
