+++ MartinyArcadeApi/Data/Models/UserRanking.cs
@@ -1,6 +1,6 @@
 public class UserRanking
 {
     public Guid UserId { get; set; }
-    public int TotalXP { get; set; }
-    public int Rank { get; set; }
+    public long TotalXP { get; set; }
+    public long Rank { get; set; }
 }
