+++ MartinyArcadeApi/Dtos/FlushXpBatchDto.cs
@@ -8,7 +8,7 @@ public class FlushXpBatchDto
 public class XpEventDto
 {
     public Guid ClientEventId { get; set; }
-    public int Amount { get; set; }
+    public long Amount { get; set; }
     public string? Reason { get; set; }
     public string Source { get; set; } = string.Empty;
 }
