+++ MartinyArcadeApi/Dtos/Leaderboard/LeaderboardEntryDto.cs
@@ -4,6 +4,6 @@ namespace MartinyArcadeApi.Dtos.Leaderboard;
 public class LeaderboardEntryDto
 {
     public string Username { get; set; } = string.Empty;
-    public int TotalXP { get; set; }
-    public int Level { get; set; }
+    public long TotalXP { get; set; }
+    public long Level { get; set; }
 }
