+++ MartinyArcadeApi/Dtos/RecordGameSessionDto.cs
@@ -4,5 +4,5 @@ public class RecordGameSessionDto
 {
     public string GameKey { get; set; } = string.Empty;
     public int Score { get; set; }
-    public int XpEarned { get; set; }
+    public long XpEarned { get; set; }
 }
